+++ session.php
@@ -17,9 +17,12 @@ class session {
             if (!headers_sent()) {
                 session_start();
                 session_regenerate_id(true);
+                return true;
             } else {
                 return false;
             }
+        } else {
+            return true;
         }
     }
 
@@ -34,7 +37,7 @@ class session {
     }
 
     /*
-     * Get Session
+     * Get Single Session
      */
 
     public function get($name) {
@@ -43,7 +46,7 @@ class session {
     }
 
     /*
-     * Remove Session
+     * Remove Single Session
      */
 
     public function remove($key) {
@@ -53,12 +56,12 @@ class session {
     }
 
     /*
-     * Remove All Session
+     * Remove All Sessions
      */
 
     public function removeAll() {
         $this->start();
-        $_SESSION = array();
+        $_SESSION = [];
         if (ini_get("session.use_cookies")) {
             $params = session_get_cookie_params();
             setcookie(session_name(), '', time() - 86400, $params["path"], $params["domain"], $params["secure"], $params["httponly"]);
@@ -67,7 +70,7 @@ class session {
     }
 
     /*
-     * Remove All Session
+     * Remove All Sessions
      * Session Destory 
      */